Save notes instead of MIDI events
Before, when we recorded, we just wrote down all the MIDI note events
and their times, then sent them back. For visualization purposes, we
want to think about the recording in terms of notes where starts and
ends are remembered together and tied to a recorded note rather than
separated.