Have display manage container instead of canvas
Before, the display module was passed a canvas for drawing in each of
its functions. Now, the display module has a container element set
before any other function calls, and the passing of canvases is no
longer needed. This way, we can use multiple canvases for layering
internally and not deal with any of this logic in the index module.