Store notes list model internally
For event listeners to be able to do things like highlight a note when
it's hovered over, they need to remember the list of notes that's
currently being displayed. One way to handle this would be to create new
event listeners each time the notes change which store the model
implicitly. This would require us to keep around the event listeners,
though, to be able to clear them before adding a new one. Instead, what
we do is keep a copy of the model initially passed to us in a global
variable and have the event listeners reference this global variable.