Add drag canvas layer above note canvas
To avoid redrawing everything every frame or close to every frame, we
redraw individual notes as their state changes (hovered, selected,
etc.) without redrawing all the other notes. We can do this because the
size of the note itself is consistent and self-contained (no overlap).
When a note is dragged or extended, it has the potential to overlap with
other notes, in which case we would need to know to redraw the
overlapped notes. To mitigate this, we introduce in this commit a second
canvas that acts as a layer above where all of the notes are drawn. When
a note starts to be dragged or extended, it gets moved to this layer so
that it can move freely without worrying about redrawing notes it sits
on top of.